requirements.txt
# Core Framework
discord.py>=2.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiohttp>=3.8.0

# Utilities
python-dotenv>=1.0.0
redis>=5.0.0  # For optional API response caching

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
aioresponses>=0.7.4
black>=23.0.0
ruff>=0.1.0

# Optional Dependencies
# pygsheets>=4.0.0  # For Google Sheets integration if needed